INTERLOK-1585 Added a dev/ssh mode. INTERLOK-1585 #time 3h
Can't actually implement a noauth plugin to the sshplugin because it will
always ask for a password (Class<String> from getCredentialsType) or use
a key (Class<PublicKey> getCredentialsType).
Write a new sshdev plugin that starts up the ssh server with UserAuthNone as
the only auth factory does the trick which means crash.properties needs to
contain.
crash.sshdev.enabled=true
crash.sshdev.port=2222
if crash.sshdev.port is not specified, it listens on a random port (which is
logged).